+++ triplea/service/graph/analysis/__init__.py
@@ -3,9 +3,11 @@ from triplea.service.graph.analysis.ganalysis import (
     sorted_degree_centrality,
     show_degree_distribution,
     sorted_average_neighbor_degree,
-    average_neighbor_number,
 )
 
+from triplea.service.graph.analysis.diameter import graph_diameter
+
+from triplea.service.graph.analysis.neighbor_number import average_neighbor_number
 
 __all__ = [
     "info",
@@ -13,4 +15,5 @@ __all__ = [
     "show_degree_distribution",
     "sorted_average_neighbor_degree",
     "average_neighbor_number",
+    "graph_diameter",
 ]